refactor(runner): split trusted authority and auth base rewriting
