+++ crates/runner/src/proxy/process.rs
@@ -1127,6 +1127,7 @@ exit 42
             "body_decoding.py",
             "body_limits.py",
             "builtin_firewall_cache.py",
+            "auth_base_rewrite.py",
             "flow_metadata_keys.py",
             "generated/__init__.py",
             "generated/model_usage.py",
@@ -1134,11 +1135,11 @@ exit 42
             "matching.py",
             "mitmproxy_compat.py",
             "registry.py",
+            "request_authority.py",
             "response_streaming.py",
             "runner_flush_lifecycle.py",
             "terminal_usage.py",
             "upstream_admission.py",
-            "url_utils.py",
             "websocket_framing.py",
             "websocket_retention.py",
             "logging_utils.py",
